LayoutTests/svg/animations/animate-keySplines-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Testing correct parsing of keySplines.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
